deps/v8/test/webkit/fast/js/Promise-resolve-with-then-exception-expected.txt

Test whether Promise treats thenable correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The promise is already rejected now.
PASS rejected
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
